fix: resolve all CI failures (clippy, fmt, testdata, tests)
- Fix 23 pre-existing clippy warnings across all crates (unused imports,
  manual_unwrap_or_default, useless_conversion, unnecessary_cast,
  too_many_arguments, sort_by_key, result_large_err)
- Add testdata TLS certificates (testdata/tls/, testdata/backendtls/)
  with proper Subject Alternative Names for TLS validation tests
- Add docs/contracts/admin-api-surface.json for contract test
- Add bundled config fixtures (configs/, deploy/) for schema validation test
- Fix XdsTlsConfig Default impl to use enabled=true (match serde default)
- Update test expectations for config default changes (xds poll 25→100ms,
  graceful drain 0→30s)
- Fix test paths: contract test ../../../docs → ../../docs,
  bundled config test ../../.. → ../..

All 4 CI jobs (cargo check, cargo test, cargo clippy -D warnings, cargo fmt)
now pass locally.

+++ crates/ntgw-http/src/proxy.rs
@@ -5,10 +5,10 @@ use std::{
     time::{Duration, Instant, SystemTime},
 };
 
-use ntgw_ai::wasm_filter::WasmPluginFilter;
 use async_trait::async_trait;
 use bytes::Bytes;
 use http::Version;
+use ntgw_ai::wasm_filter::WasmPluginFilter;
 use pingora::{
     http::{RequestHeader, ResponseHeader},
     prelude::HttpPeer,
@@ -596,7 +596,7 @@ impl ProxyHttp for GatewayProxy {
         // AI Gateway post-processing
         if let Some(ref ai_filter) = self.ai_filter {
             if let Some(ai_ctx) = ctx.ai_context.take() {
-                match ai_filter.post_process(ai_ctx, &[], ctx.status as u16).await {
+                match ai_filter.post_process(ai_ctx, &[], ctx.status).await {
                     Ok(_transformed) => {
                         tracing::debug!(
                             target: "ai_gateway",
